% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Run a bootstrap on X vector.}
\usage{
bootstrap(sample, bootstrapReplicationCount)
}
\arguments{
\item{sample}{sample.}

\item{bootstrapReplicationCount}{bootstrap replication count.}
}
\value{
A two dimensional vector representing the lower and the upper bound of CI.
}
\description{
Run a bootstrap on X vector.
}
\examples{
createBand(X = rnorm(n = 10, mean = 0, sd = 1), alpha = 0.05))
createBand(X = rnorm(n = 20, mean = 5, sd = 10), alpha = 0.15))
}
