% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapOnMA1.R
\name{bootstrapOnMA1}
\alias{bootstrapOnMA1}
\title{Generates bSampleSize bootstrap samples of length sampleSize
using W with given kernel}
\usage{
bootstrapOnMA1(sourceSample, bSampleSize = 10, kernel, bandwidth = 1)
}
\description{
Generates bSampleSize bootstrap samples of length sampleSize
using W with given kernel
}
