% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAllCorHats.R
\name{computeAllCorHats}
\alias{computeAllCorHats}
\alias{computeAllRhoHats}
\title{computeAllCorHats}
\usage{
computeAllCorHats(tParArray, lagCount, sample, kernel, bandwidth = 1)
}
\arguments{
\item{tParArray}{a T parameter array, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{lagCount}{lag.}

\item{sample}{sample.}

\item{kernel}{kernel.}

\item{bandwidth}{bandwidth.}
}
\value{

}
\description{
Computing correlations hat, that is, all \eqn{\rho} hats.
}
\examples{
 mockCorHats = computeAllCorHats(tParArray = mockTParArray,
 lagCount = myLagCount,sample = mockSample,kernel = kernel,bandwidth = 1)

}
