% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeB.R
\name{computeB}
\alias{computeB}
\alias{createB}
\title{computeB}
\usage{
computeB(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
A scalar value of \code{b}.
}
\description{
Create B, a.k.a. bandwidth given \code{n} as being a sample size
}
\examples{

}
