% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBetaLRVHat.R
\name{computeBetaLRVHat}
\alias{computeBetaLRVHat}
\alias{computeGammaHat}
\title{computeBetaLRVHat}
\usage{
computeBetaLRVHat(tParArray, lag, sample, kernel, bandwidth, allCorHats, ...)
}
\arguments{
\item{tParArray}{a T parameter array, a.k.a. a sleeper parameter, a sequence of numbers {0, 1}.}

\item{lag}{a lag value.}

\item{sample}{a sample.}

\item{kernel}{a kernel function.}

\item{bandwidth}{by default, bandwidth = 1.}
}
\value{
A scalar value of a \code{Beta LRV Hat} given certain \code{lag}.
}
\description{
Computing a \code{Beta LRV Hat} given certain \code{lag}.
}
\examples{

}
