% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIsCovered.R
\name{computeIsCovered}
\alias{computeIsCovered}
\alias{}
\title{computeIsCovered}
\usage{
computeIsCovered(band, rho)
}
\arguments{
\item{band}{band}

\item{rho}{rho}
}
\value{
Vector of {0, 1 values}.
}
\description{

}
\examples{

}
