% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIsCovered.R
\name{computeIsCovered}
\alias{computeIsCovered}
\alias{}
\title{computeIsCovered}
\usage{
computeIsCovered(band, corArray, fileName)
}
\arguments{
\item{band}{band}

\item{corArray}{corArray}
}
\value{
Vector of {0, 1 values}.
}
\description{

}
\examples{
tParCount = 10
tParArray <- createTParArray(tParCount = tParCount)
noise <- createNoise(sampleSize = tParCount, mean = 0, sd = 1)
tvMA1CoefArray <- createTVMA1CoefArray(coefFunction = sin, tParArray = tParArray)
mockTVMA1 <- createTVMA1(sampleSize = tParCount,tvMA1CoefArray = mockTVMA1CoeffArray,noise = mockNoise)
corArray <- computeCor(lag = lag, tvMa1CoefArray = tvMA1CoefArray)
lag  <- 2
band <- createBand(X = rnorm(n = 10, mean = 0, sd = 1), alpha = 0.05)
isCovered <- computeIsCovered(band = band, corArray = corArray)
}
