% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIsCoveredArray.R
\name{computeIsCoveredArray}
\alias{computeIsCoveredArray}
\alias{}
\title{computeIsCoveredArray}
\usage{
computeIsCoveredArray(bandsBrick, corArray, fileName)
}
\arguments{
\item{bandsBrick}{bandsBrick}

\item{corArray}{corArray}
}
\value{
Brick of vectors with values {0, 1}.
}
\description{
in a cycle for each band in a brick call computeIsCovered
}
