% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMEbyCovHat.R
\name{computeMEbyCovHat}
\alias{computeMEbyCovHat}
\alias{}
\title{computeMEbyCovHat}
\usage{
computeMEbyCovHat(sample, bootstrapReplicationCount)
}
\arguments{
\item{sample}{sample.}

\item{bootstrapReplicationCount}{bootstrapReplicationCount.}
}
\value{

}
\description{

}
\examples{

}
