% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageFreq.R
\name{computeNonCoverageFreq}
\alias{computeNonCoverageFreq}
\alias{}
\title{computeNonCoverageFreq}
\usage{
computeNonCoverageFreq(sampleSize, tParArray, corArray)
}
\arguments{
\item{corArray}{a tru correlation array}

\item{bandsArray}{a 3-dimensional array of confidence bands, size c(replicationCount, sampleSize, 2).}
}
\description{
For each band in bandsArray it computes isCovered and ads up all those zeros and ones. Divides this sum by replicationCount = length or dim of bandsArray
}
\examples{
computeNonCoverageFreqByCorBands(bandsArray, corArray)
}
