% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageFreqDistForLags.R
\name{computeNonCoverageFreqByRhoBands}
\alias{computeNonCoverageFreqByRhoBands}
\alias{computeNonCoverageFreqDistForLags}
\alias{}
\title{computeNonCoverageFreqByRhoBands}
\usage{
computeNonCoverageFreqDistForLags(superReplicationCount, SampleSize,
  replicationCount, nonCoverageProbability)
}
\arguments{
\item{superReplicationCount}{superReplicationCount}

\item{replicationCount}{replicationCount}

\item{nonCoverageProbability}{nonCoverageProbability}

\item{sampleSize}{sampleSize}
}
\value{

}
\description{

}
\examples{

}
