% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageProbDist.R
\name{computeNonCoverageProbDist}
\alias{computeNonCoverageProbDist}
\alias{}
\title{computeNonCoverageProbDist}
\usage{
computeNonCoverageProbDist(superReplicationCount, sampleSize, replicationCount,
  nonCoverageProbability)
}
\arguments{
\item{superReplicationCount}{superReplicationCount}

\item{sampleSize}{sampleSize}

\item{replicationCount}{replicationCount}

\item{nonCoverageProbability}{nonCoverageProbability}
}
\value{

}
\description{

}
\examples{

}
