% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageProbHat.R
\name{computeNonCoverageProbHat}
\alias{computeNonCoverageProbHat}
\alias{}
\title{computeNonCoverageProbHat}
\usage{
computeNonCoverageProbHat(sampleSize, replicationCount, nonCoverageProbability)
}
\arguments{
\item{sampleSize}{sampleSize}

\item{replicationCount}{replicationCount}

\item{nonCoverageProbability}{nonCoverageProbability}
}
\value{

}
\description{

}
\examples{

}
