% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBand.R
\name{createBand}
\alias{createBand}
\title{Create a confidence intervals band.}
\usage{
createBand(X, alpha = 0.05)
}
\arguments{
\item{X}{a vector of \code{X}.}

\item{alpha}{the probability of non-covering the real values with the estimated ones. \code{alpha = 0.05} by default.}
}
\value{
A two dimensional vector representing the lower and the upper bound of CI.
}
\description{
Create a confidence intervals band.
}
\examples{
createBand(X = rnorm(n = 10, mean = 0, sd = 1), alpha = 0.05))
}
