% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBandsBrick.R
\name{createBandsArray}
\alias{createBandsArray}
\alias{createBandsBrick}
\alias{createBands}
\title{createBandsArray}
\usage{
createBandsBrick(tParArray, lag, lagCount, bandwidth,
  kernel = normalDifferenceKernel, sampleSize,
  nonCoverageProbability = 0.05, replicationCount)
}
\arguments{
\item{nonCoverageProbability}{the probability of non-covering the real values with the estimated ones. \code{alpha = 0.05} by default.}
}
\value{
A two dimensional vector representing the lower and the upper bound of CI.
}
\description{
Create a three-dimensional array of confidence intervals bands of size c(replicationCount, sampleSize, 2).
}
\examples{

}
