% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBootstrapMultiplierR.R
\name{createBootstrapMultiplier}
\alias{createBootstrapMultiplier}
\title{Create bootstrap W series using Cholesky decomposition}
\usage{
createBootstrapMultiplier(kernel, bandwidth = 1, sampleSize)
}
\description{
Create bootstrap W series using Cholesky decomposition
}
