% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKernelCovar.R
\name{createCovarByKernel}
\alias{createCovarByKernel}
\title{Generate covariance function by given kernel}
\usage{
createCovarByKernel(kernel, bandwidth = 1, sampleSize)
}
\description{
Generate covariance function by given kernel
}
