% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCovarianceByKernel.R
\name{createCovarianceByKernel}
\alias{createCovarianceByKernel}
\title{Generate covariance function given kernel}
\usage{
createCovarianceByKernel(kernel, bandwidth = 1, sampleSize)
}
\description{
Generate covariance function given kernel
}
