% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKernelCovar.R
\name{createKernelCovar}
\alias{createKernelCovar}
\title{Generate covariance function given kernel}
\usage{
createKernelCovar(kernel, bandwidth = 1, sampleSize)
}
\description{
Generate covariance function given kernel
}
