% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiplierCovarianceByKernel.R
\name{createMultiplierCovarianceByKernel}
\alias{createMultiplierCovarianceByKernel}
\alias{createCovarianceByKernel}
\title{createMultiplierCovarianceByKernel}
\usage{
createMultiplierCovarianceByKernel(kernel, bandwidth = 1, sampleSize)
}
\arguments{
\item{kernel}{kernel.}

\item{bandwidth}{bandwidth = 1, by default}

\item{sampleSize}{the sample size.}
}
\value{
The covariance matrix.
}
\description{
Generate covariance function given kernel. Is being called from within \code{createBootstrapMultiplier} function.
}
\examples{
createMultiplierCovarianceByKernel(kernel, bandwidth = 1, sampleSize = 10)
}
