% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNoise.R
\name{createNoise}
\alias{createNoise}
\title{Create \code{noise}}
\usage{
createNoise(time, mean, sd)
}
\arguments{
\item{time}{time, a.k.a. \code{N} of the sample of \code{X}.}

\item{mean}{the mean.}

\item{sd}{standard deviation.}
}
\value{
The vector of numbers.
}
\description{
Create \code{noise}
}
\examples{
createZ(sampleSize = 10, mean = 0, sd =  1)
createZ(sampleSize = 20, mean = 5, sd =  10)
}
