% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPsi.R
\name{createPsi}
\alias{createPsi}
\title{Create psi.}
\usage{
createPsi(model, t)
}
\arguments{
\item{model}{a model, which is an underlying function used to generate \code{psi} vector.}

\item{t}{a time parameter, a sequence of numbers {0, 1}.}
}
\value{
The vector of psi values
}
\description{
Creating \code{psi} using a model. If no model provided, \code{psi} vector equals to \code{t} vector.
}
\examples{
createPsi(t =  seq(from = 0, to = 1, length.out = 10))
createPsi(t =  seq(from = 0, to = 1, by = 0.1))
Before it was not uniformly
}
