% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRho.R
\name{createRho}
\alias{createRho}
\title{Create \code{rho}}
\usage{
createRho(model, lag, sleeper)
}
\arguments{
\item{model}{the model (fixed MA1).}

\item{lag}{the lag}

\item{sleeper}{the sleeper}
}
\value{
Array \code{rho_h(i)}, counter \code{i} corresponds to \code{ t_i = 0, 1/n, …., 1 - 1/n }, horizontal ACF of a sleepers process
}
\description{
Create \code{rho}
}
\examples{
createRho(model = ma1, lag = 1, sleeper = seq(0, 1, 0.1)
}
