% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSleeper.R
\name{createSleeper}
\alias{createSleeper}
\title{Create \code{sleeper}.}
\usage{
createSleeper(number_steps)
}
\arguments{
\item{number_steps}{the desired length of \code{sleeper} vector.}
}
\value{
The vector of \code{sleeper} values
}
\description{
Create \code{sleeper}.
}
\examples{
createSleeper(number_steps = 10)
createSleeper(number_steps = 100)
}
