% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createT.R
\name{createT}
\alias{createT}
\title{Create \code{t}.}
\usage{
createT(number_steps)
}
\arguments{
\item{lengthT}{the desired length of \code{t} vector.}
}
\value{
The vector of \code{t} values
}
\description{
Create \code{t}.
}
\examples{
createT(lengthT = 10)
createT(lengthT = 100)
}
