% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTPar.R
\name{createTPar}
\alias{createTPar}
\alias{createSleeper}
\title{Create \code{T parameter}.}
\usage{
createTPar(number_steps)
}
\arguments{
\item{number_steps}{the desired length of \code{T parameter} vector.}
}
\value{
The vector of \code{sleeper} values
}
\description{
Create \code{T parameter}.
}
\examples{
createSleeper(number_steps = 10)
createSleeper(number_steps = 100)
}
