% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTVMA1.R
\name{createTVMA1}
\alias{createTVMA1}
\alias{ma1}
\title{createTVMA1}
\usage{
createTVMA1(sampleSize, tvMA1CoefArray, noise)
}
\arguments{
\item{sampleSize}{the sample size.}

\item{tvMA1CoefArray}{the \code{MA1 coefficents array}.}

\item{noise}{a white-noise vector.}
}
\value{
The vector of \code{X}
}
\description{
A manual implementation of MA1 algo.

Intended to be run as an underlying function from within \code{createX}.
Thus, supposed to be passed as the value of \code{model} option of \code{createX} function
}
\examples{
createTVMA1(8,tvMA1CoefArray,noise)
}
