% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createW.R
\name{createW}
\alias{createW}
<<<<<<< HEAD
\title{Create bootstrap W series using Cholesky decomposition of given covariance matrix}
=======
\title{Create bootstrap W series using Cholesky decomposition}
>>>>>>> 4bfb282544601b59851f24cef9e00599c5b92b58
\usage{
createW(kernel, bandwidth = 1, sampleSize)
}
\description{
<<<<<<< HEAD
Create bootstrap W series using Cholesky decomposition of given covariance matrix
=======
Create bootstrap W series using Cholesky decomposition
>>>>>>> 4bfb282544601b59851f24cef9e00599c5b92b58
}
