% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createX.R
\name{createX}
\alias{createX}
\title{Create vector of X.}
\usage{
createX(model, sampleSize, Z, psi)
}
\arguments{
\item{model}{a model, which is an underlying function used to generate X vector. If model is missing, MA(1) is going to be used.}

\item{sampleSize}{the sample size.}

\item{Z}{a white-noise vector.}

\item{psi}{the psi value.}
}
\value{
The vector of \code{X}
}
\description{
Intended to be run as an underlying function from within \code{createX}. Thus, supposed to be passed as the value of \code{model} option of \code{createX} function
}
\examples{


}
