% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma1.R
\name{ma1}
\alias{ma1}
\title{A manual implementation of MA1 algo.}
\usage{
ma1(sampleSize, psi, noise)
}
\arguments{
\item{sampleSize}{the sample size.}

\item{psi}{the \code{psi} value.}

\item{Z}{a white-noise vector.}
}
\value{
The vector of \code{X}
}
\description{
Intended to be run as an underlying function from within \code{createX}.
Thus, supposed to be passed as the value of \code{model} option of \code{createX} function
}
\examples{


}
