% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageFreqByRhoBands.R,
%   R/computeNonCoverageFreqDistForLags.R
\name{computeNonCoverageFreqByRhoBands}
\alias{computeNonCoverageFreqByRhoBands}
\alias{}
\alias{computeNonCoverageFreqByRhoBands}
\alias{computeNonCoverageFreqDistForLags}
\alias{}
\title{computeNonCoverageFreqByRhoBands}
\usage{
computeNonCoverageFreqByRhoBands(bandsArray, rho)

computeNonCoverageFreqDistForLags(superReplicationCount, SampleSize,
  replicationCount, nonCoverageProbability)
}
\arguments{
\item{bandsArray}{bandsArray.}

\item{rho}{rho}

\item{superReplicationCount}{superReplicationCount}

\item{replicationCount}{replicationCount}

\item{nonCoverageProbability}{nonCoverageProbability}

\item{sampleSize}{sampleSize}
}
\value{



}
\description{



}
\examples{


}
