% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNonCoverageFreqArray.R
\name{computeNonCoverageProbDist}
\alias{computeNonCoverageProbDist}
\alias{computeNonCoverageFreqArray}
\alias{}
\title{computeNonCoverageFreqDist}
\usage{
computeNonCoverageFreqArray(superReplicationCount, replicationCount, sampleSize,
  lag, lagCount, tParArray, kernel = normalDifferenceKernel, bandwidth,
  nonCoverageProbability)
}
\arguments{
\item{superReplicationCount}{superReplicationCount}

\item{replicationCount}{a replication count.}

\item{sampleSize}{a sample size.}

\item{lag}{a lag value, true lag is lag - 1.}

\item{tParArray}{a T parameter array}

\item{kernel}{the Epanechnikov kernel.}

\item{bandwidth}{a bandwidth value.}

\item{nonCoverageProbability}{probability of non-coverage.}
}
\value{

}
\description{

}
\examples{

}
