% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBootstrapMultiplier.R
\name{createBootstrapMultiplier}
\alias{createBootstrapMultiplier}
\alias{createBM}
\title{createBootstrapMultiplier}
\usage{
createBootstrapMultiplier(kernel, bandwidth = 1, sampleSize)
}
\arguments{
\item{kernel}{kernel.}

\item{bandwidth}{bandwidth = 1, by default}

\item{sampleSize}{the sample size.}
}
\value{
The covariance matrix.
}
\description{
Create bootstrap W series using Cholesky decomposition
}
\examples{
createBootstrapMultiplier(kernel, bandwidth = 1, sampleSize = 10)
}
