% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSample.R
\name{createSample}
\alias{createSample}
\title{Create \code{sample}.}
\usage{
createSample(model, time, noise, psi)
}
\arguments{
\item{model}{a model, which is an underlying function used to generate X vector. If model is missing, MA(1) is going to be used.}

\item{noise}{a white-noise vector.}

\item{psi}{the psi value.}

\item{sampleSize}{the sample size.}
}
\value{
The vector of \code{X}
}
\description{
Creating \code{sample}.
}
\examples{


}
