% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSample.R
\name{createSample}
\alias{createSample}
\alias{createX}
\title{createSample}
\usage{
createSample(model = createMa1, tvMA1CoefArray, noise)
}
\arguments{
\item{model}{a model, which is an underlying function used to generate a sample vector. If model is missing, MA(1) is going to be used.}

\item{tvMA1CoefArray}{the psi value.}

\item{noise}{a white-noise vector.}
}
\value{
The vector of \code{X}
}
\description{
Creating \code{sample} given \code{tvMA1CoefArray}, \code{noise} and \code{model} being a function. If no model passed, \code{ma1} is used instead.
}
\examples{
createSample(model = createMa1,
             tvMA1CoefArray = seq(from = 0,
                                  to = 1,
                                  length.out = 10),
             noise = rnorm(n = 10,
                           mean = 0,
                           sd = 1)))
}
