% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createW.R
\name{createW}
\alias{createW}
\title{Create bootstrap W series using Cholesky decomposition of given covariance matrix}
\usage{
createW(kernel, bandwidth = 1, sampleSize)
}
\description{
Create bootstrap W series using Cholesky decomposition of given covariance matrix
}
