% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createZ.R
\name{createZ}
\alias{createZ}
\title{Create Z.}
\usage{
createZ(sampleSize, mean, sd)
}
\arguments{
\item{sampleSize}{sample size, a.k.a. \code{N} of the sample of \code{X}.}

\item{mean}{the mean.}

\item{sd}{standard deviation.}
}
\value{
The vector of numbers.
}
\description{
Create Z.
}
\examples{
createZ(sampleSize = 10, mean = 0, sd =  1)
createZ(sampleSize = 20, mean = 5, sd =  10)
}
